+++ features/bootstrap/I18n/lang/pt_PT/default.php
@@ -5,13 +5,14 @@
  */
 
 return [
-    '' => array(
+    '' => [
         'plural_forms' => 'nplurals=2; plural=n!=1;'
-    ),
+    ],
 
     // Singular form
     'Hello world' => 'Olá mundo',
 
     // Plural form
-    'User' =>  array('Utilizador', 'Utilizadores')
-];;
+    'User' => ['Utilizador', 'Utilizadores'],
+    'Users' => ''
+];