+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All Notable changes to `Slick/i18n` will be documented in this file.
 
+## 1.2.0 - 2015-20-22
+
+### Added
+- Multiple domain on translation
+- Multiple locales on translation
+- Set the translator types as Translator class constants:
+  TYPE_PHP_ARRAY and TYPE_GETTEXT
+- Set class property names convention to PSR-2
+
+### removed
+- Slick/Configuration dependency
+
 ## 1.1.0 - 2015-04-14
 
 ### Added